keyboard_api/requirements.txt
fastapi
openai
uvicorn
pydantic
python-dotenv